client::component: renamed header => Header
