+++ client/src/app/project/project-list/index.project.list.js
@@ -2,7 +2,7 @@ import React, { Fragment, useEffect, useCallback } from "react";
 import { connect } from "react-redux";
 
 import Loader from "../../../components/Loader/Loader";
-import Header from "../../../components/header/header";
+import Header from "../../../components/Header/Header";
 import { projectList } from "../../../store/actions/index.action";
 import ProjectList from "../../../components/Project/ProjectList";
